+++ NEWS.md
@@ -3,8 +3,9 @@
 ## BirdFlowPipeline 0.0.9001
 Feb 9, 2024
 
-* Added NEWS.md
-* Added schema outlining major package functions, how they call each other,
+* Add NEWS.md
+* Add schema outlining major package functions, how they call each other,
 and what files they write.
 
+* Add `gpu_ram()` to calculate the GB of GPU ram needed to fit a model